update add new collection button

+++ app/src/main/java/com/nguyenminhkhang/taskmanagement/database/entity/TaskCollection.kt
@@ -7,9 +7,10 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "task_collection")
 data class TaskCollection(
     @PrimaryKey(autoGenerate = true)
-    val id: Int? = null,
+    val id: Long? = null,
     @ColumnInfo(name = "content")
     val content: String,
     @ColumnInfo(name = "updated_at")
     val updatedAt: Long
 )
+
